Add Job Options menu to Jobs list
