+++ resources/js/components/Job/StatusBadge.vue
