+++ resources/js/screens/jobs/item.vue
@@ -15,16 +15,7 @@
                         <time v-text="getCreatedAt(job)"></time>
                     </span>
                     <job-status-badge class="mr-3" :status="getStatus(job)"></job-status-badge>
-                    <div class="relative mr-3">
-                        <button :class="getOptionsButtonClass()" v-on:click="toggleOptionsMenu()">
-                            <svg aria-label="Show options" viewBox="0 0 13 16" version="1.1" width="13" height="16" role="img">
-                                <path fill-rule="evenodd" d="M1.5 9a1.5 1.5 0 1 0 0-3 1.5 1.5 0 0 0 0 3zm5 0a1.5 1.5 0 1 0 0-3 1.5 1.5 0 0 0 0 3zM13 7.5a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0z"></path>
-                            </svg>
-                        </button>
-                        <div class="absolute bg-white border rounded right-0 shadow text-sm whitespace-no-wrap" v-if="isMenuOpened">
-                            <button class="hover:bg-red-300 py-2 px-3" v-on:click="confirmDeleteJob()">Delete Job</button>
-                        </div>
-                    </div>
+                    <job-options-menu class="mr-3" :job="job"></job-options-menu>
                 </div>
             </div>
 
@@ -38,18 +29,18 @@
 <script>
     import AnsiConverter from 'ansi-to-html';
     import moment from 'moment';
-    import Swal from 'sweetalert2';
-    import JobStatusBadge from '../../components/Job/Status/Badge';
+    import JobStatusBadge from '../../components/Job/StatusBadge';
+    import JobOptionsMenu from '../../components/Job/OptionsMenu';
 
     export default {
         components: {
             JobStatusBadge,
+            JobOptionsMenu,
         },
 
         data() {
             return {
                 job: {},
-                isMenuOpened: false,
             };
         },
 
@@ -76,36 +67,6 @@
                 this.job.process.output = this.asHtml(this.job.process.output);
             },
 
-            async deleteJob() {
-                await this.$store.dispatch('deleteJobs', {
-                    id: this.$route.params.id,
-                });
-
-                await Swal.fire({
-                    type: 'success',
-                    title: 'Job Deleted!',
-                });
-
-                window.location.href = this.$store.getters.getUrl('/jobs');
-            },
-
-            async confirmDeleteJob() {
-                this.toggleOptionsMenu();
-
-                const response = await Swal.fire({
-                    type: 'warning',
-                    title: 'Job Deletion',
-                    text: 'Do you want to delete job permanently?',
-                    showCancelButton: true,
-                    confirmButtonText: 'Delete',
-                    confirmButtonColor: '#f56565',
-                });
-
-                if (response.value) {
-                    this.deleteJob();
-                }
-            },
-
             asHtml(string) {
                 return new AnsiConverter()
                     .toHtml(string)
@@ -130,16 +91,6 @@
                 } = job;
 
                 return status;
-            },
-
-            getOptionsButtonClass() {
-                const classes = 'h-full px-2 hover:bg-gray-200 rounded-full';
-
-                return classes + ` ${this.isMenuOpened ? 'bg-gray-200' : ''}`;
-            },
-
-            toggleOptionsMenu() {
-                this.isMenuOpened = !this.isMenuOpened;
             },
         },
     }
